+++ spec/support/recognition_test_case.rb
@@ -7,15 +7,13 @@ class RecognitionTestCase
   REQUEST_METHOD = "REQUEST_METHOD"
 
   def self.endpoint(body)
-    ->(env) { [200, { HEADER_ENV => env }, [body]] }
+    ->(env) { [200, {HEADER_ENV => env}, [body]] }
   end
 
   def initialize(router)
     @router = router
   end
 
-  # rubocop:disable Metrics/MethodLength
-  # rubocop:disable Metrics/AbcSize
   def run!(tests)
     tests.each do |(name, request, params, verb)|
       env                   = Rack::MockRequest.env_for(request)
@@ -32,6 +30,4 @@ class RecognitionTestCase
       end
     end
   end
-  # rubocop:enable Metrics/AbcSize
-  # rubocop:enable Metrics/MethodLength
 end