+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Hanami::Router
 Rack compatible HTTP router for Ruby
 
+## v2.0.0.alpha3 - 2020-05-20
+### Fixed
+- [Luca Guidi] `Hanami::Router#initialize` do not yield block if not given
+- [Luca Guidi] Ensure to not accidentally cache response headers for HTTP 404 and 405
+- [Luca Guidi] Ensure scoped root to not be added as trailing slash
+
+## v2.0.0.alpha2 - 2020-02-19
+### Added
+- [Luca Guidi] Block syntax. Routes definition accept a block which returning value is the body of the Rack response.
+- [Luca Guidi] Added `resolver:` option to `Hanami::Router#initialize` to provide your own strategy to load endpoints.
+
+### Changed
+- [Luca Guidi] Removed `Hanami::Router#resource` and `#resources`.
+- [Luca Guidi] Removed loading of routes endpoints.
+- [Luca Guidi] Removed `inflector:` from `Hanami::Router#initialize`
+- [Luca Guidi] Removed `scheme:`, `host:`, `port:` from `Hanami::Router#initialize`, use `base_url:` instead.
+
 ## v2.0.0.alpha1 - 2019-01-30
 ### Added
 - [Luca Guidi] Introduce `Hanami::Router#scope` to support single routing tier for Hanami
@@ -16,6 +33,11 @@ Rack compatible HTTP router for Ruby
 - [Luca Guidi] Router is frozen after initialization
 - [Luca Guidi] All the code base respects the frozen string pragma
 - [Luca Guidi] `Hanami::Router#initialize` requires `configuration:` option if routes endpoints are `Hanami::Action` subclasses
+
+## v1.3.2 - 2019-02-13
+### Added
+- [Luca Guidi] Official support for Ruby: MRI 2.7
+- [Luca Guidi] Support `rack` 2.1
 
 ## v1.3.1 - 2019-01-18
 ### Added